Modify tests and move to separate file
