--- test/refasm/Intel64/Windows/bytecode/exponentiation_by_squaring_0.tst
@@ -1,5 +1,0 @@
-exponentiation_by_squaring_0(i0)
-     0 : mov i1, i1 
-     1 : mov i1, 1  
-     2 : mov i0, i1 
-     3 : ret
