--- test/refasm/Intel64/Windows/bytecode/exponentiation_by_squaring_1.tst
@@ -1,6 +1,0 @@
-exponentiation_by_squaring_1(i0)
-     0 : mov i1, i1     
-     1 : mov i2, 1      
-     2 : mul i2, i2, i1 
-     3 : mov i0, i2     
-     4 : ret
