--- test/refasm/Intel64/Windows/bytecode/exponentiation_by_squaring_9.tst
@@ -1,10 +1,0 @@
-exponentiation_by_squaring_9(i0)
-     0 : mov i1, i1     
-     1 : mov i2, 1      
-     2 : mul i2, i2, i1 
-     3 : mul i1, i1, i1 
-     4 : mul i1, i1, i1 
-     5 : mul i1, i1, i1 
-     6 : mul i2, i2, i1 
-     7 : mov i0, i2     
-     8 : ret
