--- test/refasm/Intel64/Windows/exponentiation_by_squaring_0.tst
@@ -1,5 +1,0 @@
-exponentiation_by_squaring_0(i0)
-     0 : mov rcx, rcx   ; 48 89 c9              
-     1 : mov rcx, #0x01 ; 48 c7 c1 01 00 00 00  
-     2 : mov rax, rcx   ; 48 89 c8              
-     3 : ret            ; c3
