+++ test/refasm/aarch64/bytecode/exponentiation_by_squaring_0.tst
@@ -1,0 +1,5 @@
+exponentiation_by_squaring_0(i0)
+     0 : mov i0, i0 
+     1 : mov i0, 1  
+     2 : mov i0, i0 
+     3 : ret
