+++ test/refasm/aarch64/bytecode/exponentiation_by_squaring_1.tst
@@ -1,0 +1,6 @@
+exponentiation_by_squaring_1(i0)
+     0 : mov i0, i0     
+     1 : mov i1, 1      
+     2 : mul i0, i1, i0 
+     3 : mov i0, i0     
+     4 : ret
