+++ test/refasm/aarch64/bytecode/exponentiation_by_squaring_9.tst
@@ -1,0 +1,10 @@
+exponentiation_by_squaring_9(i0)
+     0 : mov i0, i0     
+     1 : mov i1, 1      
+     2 : mul i1, i1, i0 
+     3 : mul i0, i0, i0 
+     4 : mul i0, i0, i0 
+     5 : mul i0, i0, i0 
+     6 : mul i0, i1, i0 
+     7 : mov i0, i0     
+     8 : ret
