+++ test/refasm/aarch64/exponentiation_by_squaring_0.tst
@@ -1,0 +1,5 @@
+exponentiation_by_squaring_0(i0)
+     0 : mov x0, x0    ; e0 03 00 aa  
+     1 : mov x0, #0x01 ; 20 00 80 d2  
+     2 : mov x0, x0    ; e0 03 00 aa  
+     3 : ret x30       ; c0 03 5f d6
